[tast]: Add modemmanager control tests
Add to validate enable, disable, connect, disconnect dbus calls on modemamanager.
Helps to validate cellular at MM level to have clarity on shill vs MM

BUG=b:181016882, b:192296455, b:198377954
TEST=tast --verbose run localhost:2226 network.modemmanager_enable_and_connect

+++ src/chromiumos/tast/local/bundles/cros/network/shill_cellular_roaming.go
@@ -84,7 +84,7 @@ func ShillCellularRoaming(ctx context.Context, s *testing.State) {
 	}
 	defer cleanup(ctxForCleanUp, s)
 
-	if err := modem.WaitForState(ctx, mmconst.ModemStateRegistered, time.Minute); err != nil {
+	if err := modem.WaitForState(ctx, int32(mmconst.ModemStateRegistered), time.Minute); err != nil {
 		s.Fatal("Modem is not registered")
 	}
 
@@ -106,7 +106,7 @@ func ShillCellularRoaming(ctx context.Context, s *testing.State) {
 		s.Fatal("Could not set AllowRoaming property to false: ", err)
 	}
 
-	if err := modem.WaitForState(ctx, mmconst.ModemStateRegistered, time.Minute); err != nil {
+	if err := modem.WaitForState(ctx, int32(mmconst.ModemStateRegistered), time.Minute); err != nil {
 		s.Fatal("Modem is not registered")
 	}
